feat: jpa연관관계 매핑 제거 작업 완료 / CI 구축 테스트
- JPA 연관관계 매핑 제거 작업 완료. 제거 후 아래와 같은 장점을 얻을 수 있음
  - DTO Projection을 사용하여 필요한 필드만 조회하도록 변경
  - 외래키를 직접 설정할 수 있도록 함.
  - 불필요한 엔티티 조회 하지 않음. 기존: 엔티티 새롭게 저장하고, 엔티티에 연관관계가 있음으로써 연관관계 설정할때 연관관계 엔티티 조회 후 설정
  - 비즈니스 로직을 추상화 -> 구체적으로 직접 구축할 수 있음. 이에 따라 비즈니스 로직 최적화가 가능
  - 비즈니스 로직을 구체화 함으로써 일부 삭제시에 기존에는 조회후 삭제 -> 바로 삭제 처리, 업데이트시 해당 엔티티 조회 -> 업데이트 대신 바로 업데이트 처리, save시에도 외래키만 직접 설정하기 때문에 연관 엔티티 조회 후 매핑할 필요가x
  - 추가로 카테고리 부분 비즈니스 로직을 꽤나 많이 최적화할 수 있게 되었음

- JPA 연관관계 매핑 제거 후 아래와 같은 단점?이 발생
 - 성능이 나아졌지만, 비즈니스 로직을 확실하게 처리하지 않으면 데이터 정합성에 문제가 발생할 수 있음
 - 연관관계 매핑을 했다면 객체 그래프 탐색으로 쉽게 처리할 수 있는 장점이 있지만(생산성 향상), 성능의 차이가 크다고 판단하여 매핑 제거

- 게시글 조회 시 게시글 데이터 / 게시글 총 개수 2개의 쿼리로 분리
 - 이유는 게시글 총 개수의 경우 인덱스 트리에 있는 총 개수를 조회함으로써 빠르게 조회할 수 있음 - 단점은 쿼리가 1번 더 발생
- 카테고리 같은 경우 부모(자식) 카테고리에 연결된 게시글 수 총합 가져오는 서브쿼리까지 하나의 쿼리로 묶음
 - 카테고리 데이터 / 카테고리와 연결된 게시글 수 총합을 따로 분리해서 처리하면 부모(자식)카테고리 데이터, 부모(자식)의 카테고리와 연결된 게시글 총 개수 총 4개의 쿼리가 발생하게 되어 비효율적이라고 판단

- 결론적으로 위와 같이 불필요한 쿼리 최소화, 비즈니스 로직 최적화, 필요한 필드만 조회 등을 함으로써 성능이 많이 좋아졌음. 예를 들어 게시글 1000개 기준 4000ms -> 10ms로 성능 향상

- 위 로직을 수정함에 따라 테스트 코드 수정 완료, 프론트와 연계하여 오류가 없는지 확인하였고 오류가 발생한 부분 수정 완료

- CI 구축 첫 번째 테스트 커밋

+++ blog-springboot-jpa/src/main/java/com/yhs/blog/springboot/jpa/domain/post/service/impl/PostOperationServiceImpl.java
@@ -100,7 +100,7 @@ public class PostOperationServiceImpl implements PostOperationService {
             fileService.processCreateFilesForCreatePostRequest(postRequest.getFiles(), savedPost.getId());
 
             // 태그 처리
-            processPostTagsAndTags(post, postRequest.getTags(), savedPost.getId());
+            processPostTagsAndTags(postRequest.getTags(), savedPost.getId());
 
             log.info("[PostOperationServiceImpl] createNewPost S3 메인 스레드 시작: {}", Thread.currentThread().getName());
 
@@ -145,6 +145,8 @@ public class PostOperationServiceImpl implements PostOperationService {
 
         try {
 
+            // 아래 수정할때 조회 -> 수정 대신 바로 게시글 수정할수도 있는데, 조회할때 찾지 못하면 예외 발생 시켜야 하기 때문에 조회 -> 수정
+            // 처리
             // fetch join으로 user까지 함께 가져와 영속성 컨텍스트에 저장. 이후 getUser()로 가져올때 추가 select 쿼리 없음
             Post post = postRepository.findById(postId)
                     .orElseThrow(() -> new BusinessException(ErrorCode.POST_NOT_FOUND, postId +
@@ -164,7 +166,7 @@ public class PostOperationServiceImpl implements PostOperationService {
                 categoryId = null;
             }
 
-            // 파일(이미지) 삭제 및 저장 처리
+            // 파일(이미지) 삭제 및 저장 처리 - 대표 이미지 포함
             fileService.processUpdateFilesForUpdatePostRequest(postUpdateRequest.getFiles(), postId,
                     postUpdateRequest.getDeletedImageUrlsInFuture());
 
@@ -184,9 +186,10 @@ public class PostOperationServiceImpl implements PostOperationService {
                 tagRepository.deleteAll(unusedTags);
             }
 
-            processPostTagsAndTags(post, postUpdateRequest.getTags(), postId);
+            processPostTagsAndTags(postUpdateRequest.getTags(), postId);
 
-            // 대표 이미지 처리
+            // 대표 이미지 설정, 대표 이미지 삭제는 위쪽 fileService.processUpdateFilesForUpdatePostRequest에서
+            // 일괄 처리 했음
             Long featuredImageId = featuredImageService
                     .processFeaturedImageForUpdatePostRequest(postUpdateRequest.getFeaturedImage());
 
@@ -194,7 +197,7 @@ public class PostOperationServiceImpl implements PostOperationService {
             String convertedContent = postUpdateRequest.getContent().replace(blogId + "/temp/",
                     blogId + "/final/");
 
-            // 더티 체킹 대상으로써 save 메서드 불필요
+            // 더티 체킹 대상으로써 save 메서드 불필요.
             post.update(categoryId, featuredImageId, postUpdateRequest.getTitle(), convertedContent,
                     PostStatus.valueOf(postUpdateRequest.getPostStatus().toUpperCase()),
                     CommentsEnabled.valueOf(postUpdateRequest.getCommentsEnabled().toUpperCase()));
@@ -226,26 +229,44 @@ public class PostOperationServiceImpl implements PostOperationService {
     public void deletePostByPostId(Long postId, BlogUser blogUser) {
         log.info("[PostOperationServiceImpl] deletePostByPostId 메서드 시작: postId: {}", postId);
 
-        String blogId = blogUser.getBlogIdFromToken();
-
         try {
 
+            // 아래 삭제할때 조회 -> 삭제 대신 바로 삭제할수도 있는데, 조회할때 찾지 못하면 예외 발생 시켜야 하기 때문에 조회 -> 삭제
+
+            String blogId = blogUser.getBlogIdFromToken();
+
+            Post foundPost = postRepository.findById(postId)
+                    .orElseThrow(() -> new BusinessException(ErrorCode.POST_NOT_FOUND, postId + "번 게시글을 찾을 수 없습니다.",
+                            "PostOperationServiceImpl", "deletePostByPostId"));
+
+            List<String> toBeDeletedFileUrls = new ArrayList<>();
+
             // 파일 삭제 처리
-            List<String> toBeDeletedFileUrls = fileService.processDeleteFilesForDeletePostRequest(postId);
+            List<String> fileUrls = fileService.processDeleteFilesForDeletePostRequest(postId);
+
+            if (fileUrls != null && !fileUrls.isEmpty()) {
+                toBeDeletedFileUrls.addAll(fileUrls);
+            }
 
             // 대표 이미지 삭제 처리
             Long featuredImageId = postRepository.findFeaturedImageIdByPostId(postId);
-            featuredImageService.processDeleteFeaturedImageForDeletePostRequest(featuredImageId);
+
+            if (featuredImageId != null) {
+                String featuredImageFileUrl = featuredImageService
+                        .processDeleteFeaturedImageForDeletePostRequest(featuredImageId);
+                toBeDeletedFileUrls.add(featuredImageFileUrl);
+            }
 
             // 태그 삭제 처리, PostTag는 무조건 삭제하면 됨
             postTagRepository.deletePostTagsByPostId(postId);
             List<Tag> toBeDeletedTags = tagRepository.findUnusedTagsByPostId(postId);
-            tagRepository.deleteAllInBatch(toBeDeletedTags);
+
+            if (toBeDeletedTags != null && !toBeDeletedTags.isEmpty()) {
+                tagRepository.deleteAllInBatch(toBeDeletedTags);
+            }
 
             // 게시글 삭제 처리
-            Post foundPost = postRepository.findById(postId)
-                    .orElseThrow(() -> new BusinessException(ErrorCode.POST_NOT_FOUND, postId + "번 게시글을 찾을 수 없습니다.",
-                            "PostOperationServiceImpl", "deletePostByPostId"));
+
             postRepository.delete(foundPost);
 
             // 트랜잭션이 성공되어야만 실행
@@ -266,7 +287,7 @@ public class PostOperationServiceImpl implements PostOperationService {
 
     }
 
-    private void processPostTagsAndTags(Post post, List<String> tagNames, Long postId) {
+    private void processPostTagsAndTags(List<String> tagNames, Long postId) {
 
         log.info("[PostOperationServiceImpl] processTags 메서드 시작");
 
@@ -281,15 +302,25 @@ public class PostOperationServiceImpl implements PostOperationService {
                 Optional<Tag> tag = tagRepository.findByName(tagName);
 
                 if (tag.isPresent()) {
+                    postTags.add(new PostTag(postId, tag.get().getId()));
                     continue;
                 }
 
-                tags.add(tag.get()); // 새롭게 생성된 태그를 영속성 컨텍스트에 저장
-                PostTag postTag = new PostTag(postId, tag.get().getId());
-                postTags.add(postTag);
+                Tag newTag = new Tag(tagName);
+
+                tags.add(newTag);
             }
-            tagRepository.saveAll(tags);
+
+            List<Tag> savedTags = tagRepository.saveAll(tags);
+
+            savedTags.forEach(tag -> {
+                PostTag postTag = new PostTag(postId, tag.getId());
+                postTags.add(postTag);
+            });
+
             postTagRepository.saveAll(postTags);
+        } else {
+            return;
         }
     }
